refactor(asset-designer): Separate VariableSelector
